```
| Method |     Mean |   Error |  StdDev |
|------- |---------:|--------:|--------:|
|    New | 104.2 us | 2.06 us | 3.98 us |
|    Old | 145.4 us | 2.83 us | 2.78 us |
```

```cs
    private IRef<object> _refOld = RefCountableOld.Create(new Disp());
    private IRef<object> _refNew = RefCountableNew.Create(new Disp());

    [Benchmark]
    public void New()
    {
        for(var c=0;c<1000;c++) _refNew.Clone().Dispose();
    }


    [Benchmark]
    public void Old()
    {
        for(var c=0;c<1000;c++) _refOld.Clone().Dispose();
    }
```