@kekekeks 2025-08-17
Probably because it's easier to reason about code wrapped into `lock` blocks.

We were using locks in the following operations:
1) `Clone`
The lock was used to check if the current ref is not disposed yet and then call AddRef on the shared ref-counter, so we can be sure that no other thread would dispose the current ref while we are incrementing the ref counter.
 The new code instead takes a defensive copy of the ref counter and just _tries_ to add ref. If it fails, that it means that the current ref got disposed in the meantime. There is no observable behavior change.
2) `Dispose`
The lock was used to prevent double dispose and to make sure that lock in (1) is meaningful. The new code just uses Interlocked.Exchange to make sure that there is no double dispose
3) `get_Item` -  the lock here was completely useless